Update PA EE to use tabs provider
